Imported using usd-importer.

Publish parent: fd26c6fca1173c1c885f08af0768ef46979d16ee
Changelog parent: e8fa178f44bda79f1abaef6f0061e3059f47cfff

New changelog entries:
  [ Aurelien Jarno ]
  * Update from upstream stable branch:
      - Fixes bug18240 failing with a timeout on machines with a lot of swap.
  * patches/any/cvs-grantpt-pty-owner.diff: new patch from upstream to
    improve granpt when /dev/pts is not mounted with the correct options.
  * rules.d/debhelper.mk: only install pt_chown when built.
  * sysdeps/linux.mk: don't build pt_chown (CVE-2013-2207). Closes: #717544.
  [ Aurelien Jarno ]
  * Update from upstream stable branch:
    - Fix segmentation fault caused by passing out-of-range data to strftime()
      (CVE-2015-8776).  Closes: #812445.
    - Fix an integer overflow in hcreate() and hcreate_r() (CVE-2015-8778).
      Closes: #812441.
    - Fix multiple unbounded stack allocations in catopen() (CVE-2015-8779).
      Closes: #812455.
  * patches/any/local-CVE-2015-7547.diff: new patch to fix glibc getaddrinfo
    stack-based buffer overflow (CVE-2015-7547).

+++ misc/chflags.c
@@ -1,4 +1,4 @@
-/* Copyright (C) 1991-2015 Free Software Foundation, Inc.
+/* Copyright (C) 1991-2014 Free Software Foundation, Inc.
    This file is part of the GNU C Library.
 
    The GNU C Library is free software; you can redistribute it and/or
@@ -21,12 +21,12 @@
 
 /* Change the flags of FILE to FLAGS.  */
 
-int chflags (const char *file, unsigned long int flags) __THROW;
+int chflags (const char *file, int flags) __THROW;
 
 int
 chflags (file, flags)
      const char *file;
-     unsigned long int flags;
+     int flags;
 {
   if (file == NULL)
     {
